Refactor front page to show translated components
