+++ ui/src/App.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
-import { FrontPage } from './components/FrontPage'
+import { FrontPage } from './components/FrontPage/FrontPage'
 import { FacilitySearch } from './components/FacilitySearch/FacilitySearch'
 import Navigation from './components/Navigation'
 import { NavigationItem } from './models/NavigationItem'
