+++ ui/src/components/FrontPage/FrontPageContentEn.tsx
@@ -1,0 +1,3 @@
+export const FrontPageContentEn = () => {
+  return <div>{/* Front page elements & content in English */}</div>
+}
