+++ ui/src/components/FrontPage/FrontPageContentFi.tsx
@@ -1,0 +1,3 @@
+export const FrontPageContentFi = () => {
+  return <div>{/* Front page elements & content in Finnish */}</div>
+}
