+++ ui/src/components/FrontPage/FrontPageContentSv.tsx
@@ -1,0 +1,3 @@
+export const FrontPageContentSv = () => {
+  return <div>{/* Front page elements & content in Swedish */}</div>
+}
